refactor(api): extract zod schemas to dedicated module for Next.js 15 compat
