+++ src/app/api/calendario-academico/[id]/eventos/route.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 import { getContainer } from "@/lib/server/container";
 import { ApiError, fromZodError } from "@/lib/server/errors";
 import { withAdmin } from "@/lib/server/services/auth/middleware";
-import { ALL_CATEGORIAS } from "@/lib/shared/config/calendario-academico";
+import { createEventoSchema } from "@/lib/server/api-schemas/calendario";
 
 type RouteContext = { params: Promise<{ id: string }> };
 
@@ -23,18 +23,6 @@ export async function GET(_request: Request, context: RouteContext) {
     return ApiError.internal("Erro ao buscar eventos");
   }
 }
-
-const dateString = z
-  .string()
-  .min(1, "Data é obrigatória")
-  .refine(s => !isNaN(new Date(s).getTime()), { message: "Data inválida" });
-
-export const createEventoSchema = z.object({
-  descricao: z.string().min(1, "Descrição é obrigatória"),
-  dataInicio: dateString,
-  dataFim: dateString,
-  categoria: z.enum(ALL_CATEGORIAS).default("OUTRA"),
-});
 
 export function POST(request: Request, context: RouteContext) {
   return withAdmin(request, async req => {
