+++ src/app/api/usuarios/me/disciplinas/marcar/route.ts
@@ -2,14 +2,9 @@ import { NextResponse } from "next/server";
 import { withAuth } from "@/lib/server/services/auth/middleware";
 import { ApiError } from "@/lib/server/errors";
 import { getContainer } from "@/lib/server/container";
-import { z } from "zod";
+import { marcarDisciplinasSchema } from "@/lib/server/api-schemas/usuarios";
 
 export const dynamic = "force-dynamic";
-
-export const marcarDisciplinasSchema = z.object({
-  disciplinaIds: z.array(z.string().uuid()).min(1),
-  status: z.enum(["concluida", "cursando", "none"]),
-});
 
 /**
  * POST /api/usuarios/me/disciplinas/marcar
