+++ src/app/api/usuarios/me/membros/[membroId]/route.ts
@@ -4,21 +4,11 @@ import { z } from "zod";
 import { withAuth } from "@/lib/server/services/auth/middleware";
 import { getContainer } from "@/lib/server/container";
 import { ApiError, fromZodError } from "@/lib/server/errors";
+import { updateOwnMembershipSchema } from "@/lib/server/api-schemas/usuarios";
 
 type RouteContext = {
   params: Promise<{ membroId: string }>;
 };
-
-const dateStringSchema = z
-  .string()
-  .refine(v => !isNaN(Date.parse(v)), { message: "Data inválida" });
-
-export const updateOwnMembershipSchema = z.object({
-  papel: z.enum(["ADMIN", "MEMBRO"]).optional(),
-  cargoId: z.string().uuid("ID de cargo inválido").nullable().optional(),
-  startedAt: dateStringSchema.optional(),
-  endedAt: dateStringSchema.nullable().optional(),
-});
 
 export async function PUT(request: Request, context: RouteContext) {
   return await withAuth(request, async (req, usuario) => {
