+++ src/app/api/usuarios/me/onboarding/route.ts
@@ -2,37 +2,9 @@ import { NextResponse } from "next/server";
 import { withAuth } from "@/lib/server/services/auth/middleware";
 import { getContainer } from "@/lib/server/container";
 import { handleError } from "@/lib/server/errors";
-import { z } from "zod";
+import { onboardingPatchSchema } from "@/lib/server/api-schemas/usuarios";
 
 export const dynamic = "force-dynamic";
-
-const stepStateSchema = z
-  .object({
-    completedAt: z.string().optional(),
-    skippedAt: z.string().optional(),
-  })
-  .strict();
-
-export const patchSchema = z
-  .object({
-    welcome: z.object({ completedAt: z.string() }).strict().optional(),
-    periodo: stepStateSchema.optional(),
-    concluidas: stepStateSchema.optional(),
-    entidades: stepStateSchema.optional(),
-    done: z.object({ completedAt: z.string() }).strict().optional(),
-    semesters: z
-      .record(
-        z.string(),
-        z
-          .object({
-            cursando: stepStateSchema.optional(),
-            turmas: stepStateSchema.optional(),
-          })
-          .strict()
-      )
-      .optional(),
-  })
-  .strict();
 
 export function GET(request: Request) {
   return withAuth(request, async (_req, usuario) => {
@@ -50,7 +22,7 @@ export function PATCH(request: Request) {
   return withAuth(request, async (req, usuario) => {
     try {
       const body = await req.json();
-      const parsed = patchSchema.parse(body);
+      const parsed = onboardingPatchSchema.parse(body);
       const { usuariosRepository } = getContainer();
       await usuariosRepository.updateOnboardingMetadata(usuario.id, parsed);
       const updated = await usuariosRepository.getOnboardingMetadata(usuario.id);
